refactor(icon): unify fileName with componentName in PascalCase
