+++ packages/icon/src/components/CheckCircleIcon.tsx
