+++ packages/icon/src/components/EmailIcon.tsx
