+++ packages/icon/src/components/KakaoIcon.tsx
