+++ packages/icon/src/components/YoutubeIcon.tsx
